refactoring & authService & tests

+++ src/main/resources/db/migration/V10002__google_sheet.sql
@@ -3,6 +3,6 @@ create table if not exists "google_sheet_tb"
 (
     id bigint default nextval('seq_google_sheet') not null primary key,
     table_id bigint not null,
-    sheet_id bigint not null,
+    sheet_id int not null,
     sheet_name varchar(200) not null
 );
